+++ components/Menu.vue
@@ -122,11 +122,11 @@
       ...mapGetters(['categories']),
 
       nav() {
-        return this.$store.state.nav;
+        return this.$store.state;
       },
 
       showContacts() {
-        return this.$store.state.firebase.showContacts;
+        return this.$store.state.showContacts;
       },
     },
