+++ layouts/home.vue
@@ -107,7 +107,7 @@
 
     computed: {
       nav() {
-        return this.$store.state.nav;
+        return this.$store.state;
       }
     },
